project/web/flask_socketio/requirements.txt
Flask==1.1.2
Flask-gzip==0.2
Flask-SocketIO==4.3.2
